Added security to views

+++ products/views.py
@@ -2,12 +2,11 @@ from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
 from django.db.models.functions import Lower
-
+from django.contrib.auth.decorators import login_required
 from .models import Product, Category
 from .forms import ProductForm
 
 
-# Create your views here.
 
 def all_products(request):
     """ A view to show all cakes, including sorting """
@@ -70,8 +69,12 @@ def product_detail(request, product_id):
     return render(request, 'products/cake_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """ Add a cake to the  store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permission to do that!.')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -92,8 +95,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a cake in the website """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permission to do that!.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -117,8 +125,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a cake in the website """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, you do not have permission to do that!.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Cake deleted!')
